# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'MD Global'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  /api/v1/medical stuff/help-and-support:
    post:
      summary: 'Help and Support'
      operationId: helpAndSupport
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Medical Stuff Settings'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                  description: ''
                  example: qui
                message:
                  type: string
                  description: "@response 200{\n\"status\": true,\n\"data\": {\n\"id\": 1,\n\"user_id\": 1,\n\"name\": \"test medical stuff\",\n\"email\": \"test@yopmail.com\",\n\"phone\": \"1234567890\",\n\"subject\": \"test\",\n\"message\": \"test\",\n\"created_at\": \"2021-09-28T10:20:00.000000Z\",\n\"updated_at\": \"2021-09-28T10:20:00.000000Z\"\n}\n}"
                  example: eum
              required:
                - subject
                - message
      security: []
...